hisi-idt.py: fix "failed" messages
The acknowledge detection in bootdownload.sendframe() is incorrect,
which causes "failed" messages and a longer than usual flash time
(the transfer is actually successful):

 $ hisi-idt.py --img1=../l-loader/recovery.bin
 +----------------------+
  Serial:  /dev/serial/by-id/usb-䕇䕎䥎_㄰㌲㔴㜶㤸-if00-port0
  Image1:  ../l-loader/recovery.bin
 +----------------------+

 Sending ../l-loader/recovery.bin ...
 failed
 failed
 failed
 failed
 failed
 failed
 failed
 failed
 failed
 failed
 failed
 Done

Obviously, Python3 does not want a chr() here; a byte value has to be
used instead. In addition, then length check seems totally redundant so
remove it.

Signed-off-by: Jerome Forissier <jerome@forissier.org>