hisi-idt.py: retry connection on ENOENT and EBUSY
I observed the following behavior with my HiKey board:

- If I run hisi-idt.py less than 1 second after powering up the board,
the device is not found under /dev (errno == ENOENT);
- About 2 seconds after power up, hisi-idt.py succeeds;
- Between 3 and 25 seconds after power up, the connection fails with
errno == EBUSY;
- Between 25 and 75 seconds, flashing works fine;
- More than 75 seconds after powerup, the device is gone.

I have no idea what causes the EBUSY error, but it is annoying enough
to do something about it. Especially since flashing in recovery mode is
not a very common task, so the errors can get you on a wrong track
wondering if something has been broken.

This commit introduces a retry mechanism to make the operation more
reliable. In practice, hisi-idt.py can now be started before or after
the board is powered up.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Tested-by: Joakim Bech <joakim.bech@linaro.org> (Hikey 6220)